requirements.txt
Django==1.9.7
django-bootstrap3==7.0.1
django-filter==0.13.0
djangorestframework==3.3.3
Markdown==2.6.6
requests==2.10.0
